move collection functions into utils
