+++ siliconcompiler/project.py
@@ -2,7 +2,6 @@ import importlib
 import inspect
 import logging
 import os
-import shutil
 import sys
 import tarfile
 import uuid
@@ -14,7 +13,6 @@ from typing import Set, Union, List, Tuple, Type, Callable, TextIO
 
 from siliconcompiler.schema import BaseSchema, NamedSchema, EditableSchema, Parameter, Scope, \
     __version__ as schema_version
-from siliconcompiler.schema.parametervalue import NodeListValue, NodeSetValue
 from siliconcompiler.schema.utils import trim
 
 from siliconcompiler import Design
@@ -35,7 +33,7 @@ from siliconcompiler.schema_support.pathschema import PathSchemaBase
 from siliconcompiler.report.dashboard.cli import CliDashboard
 from siliconcompiler.scheduler import Scheduler, SchedulerNode
 from siliconcompiler.utils.logging import SCColorLoggerFormatter, SCLoggerFormatter
-from siliconcompiler.utils import FilterDirectories, get_file_ext
+from siliconcompiler.utils import get_file_ext
 from siliconcompiler.utils.multiprocessing import MPManager
 from siliconcompiler.flows.showflow import ShowFlow
 from siliconcompiler.flowgraph import RuntimeFlowgraph
@@ -756,166 +754,6 @@ class Project(PathSchemaBase, CommandLineSchema, BaseSchema):
 
             dirlist.append(str(index))
         return os.path.join(*dirlist)
-
-    def getcollectiondir(self) -> str:
-        """
-        Returns the absolute path to the directory where collected files are stored.
-
-        This directory is typically located within the project's working directory
-        and is used to consolidate files marked for collection.
-
-        Returns:
-            str: The absolute path to the collected files directory.
-        """
-        return os.path.join(self.getworkdir(), "sc_collected_files")
-
-    def collect(self,
-                directory: str = None,
-                verbose: bool = True,
-                whitelist: List[str] = None):
-        '''
-        Collects files and directories specified in the schema and places
-        them in a collection directory. The function only copies items that have
-        the 'copy' field set to True in their schema definition.
-
-        Args:
-            directory (str, optional): The output directory for collected files.
-                Defaults to the path from :meth:`.getcollectiondir`.
-            verbose (bool): If True, logs information about each collected file/directory.
-                Defaults to True.
-            whitelist (List[str], optional): A list of absolute paths that are
-                allowed to be collected. If an item to be collected is not on this list,
-                a `RuntimeError` is raised. Defaults to None.
-
-        Raises:
-            RuntimeError: If a file or directory to be collected is not in the `whitelist`.
-            FileNotFoundError: If a specified file or directory cannot be found.
-        '''
-        if not directory:
-            directory = self.getcollectiondir()
-        directory = os.path.abspath(directory)
-
-        # Remove existing directory
-        if os.path.exists(directory):
-            shutil.rmtree(directory)
-        os.makedirs(directory)
-
-        if verbose:
-            self.logger.info(f'Collecting files to: {directory}')
-
-        dirs = {}
-        files = {}
-
-        for key in self.allkeys():
-            if key[0] == 'history':
-                # skip history
-                continue
-
-            # Skip runtime directories
-            if key == ('option', 'builddir'):
-                # skip builddir
-                continue
-            if key == ('option', 'cachedir'):
-                # skip cache
-                continue
-
-            if key[0] == 'tool' and key[2] == 'task' and key[4] in ('input',
-                                                                    'report',
-                                                                    'output'):
-                # skip flow files files from builds
-                continue
-
-            leaftype = self.get(*key, field='type')
-            is_dir = "dir" in leaftype
-            is_file = "file" in leaftype
-
-            if not is_dir and not is_file:
-                continue
-
-            if not self.get(*key, field='copy'):
-                continue
-
-            for values, step, index in self.get(*key, field=None).getvalues(return_values=False):
-                if not values.has_value:
-                    continue
-
-                if isinstance(values, (NodeSetValue, NodeListValue)):
-                    values = values.values
-                else:
-                    values = [values]
-
-                if is_dir:
-                    dirs[(key, step, index)] = values
-                else:
-                    files[(key, step, index)] = values
-
-        path_filter = FilterDirectories(self)
-        for key, step, index in sorted(dirs.keys()):
-            abs_paths = self.find_files(*key, step=step, index=index)
-
-            new_paths = set()
-
-            if not isinstance(abs_paths, (list, tuple, set)):
-                abs_paths = [abs_paths]
-
-            abs_paths = zip(abs_paths, dirs[(key, step, index)])
-            abs_paths = sorted(abs_paths, key=lambda p: p[0])
-
-            for abs_path, value in abs_paths:
-                if not abs_path:
-                    raise FileNotFoundError(f"{value.get()} could not be copied")
-
-                if abs_path.startswith(directory):
-                    # File already imported in directory
-                    continue
-
-                imported = False
-                for new_path in new_paths:
-                    if abs_path.startwith(new_path):
-                        imported = True
-                        break
-                if imported:
-                    continue
-
-                new_paths.add(abs_path)
-
-                import_path = os.path.join(directory, value.get_hashed_filename())
-                if os.path.exists(import_path):
-                    continue
-
-                if whitelist is not None and abs_path not in whitelist:
-                    raise RuntimeError(f'{abs_path} is not on the approved collection list.')
-
-                if verbose:
-                    self.logger.info(f"  Collecting directory: {abs_path}")
-                path_filter.abspath = abs_path
-                shutil.copytree(abs_path, import_path, ignore=path_filter.filter)
-                path_filter.abspath = None
-
-        for key, step, index in sorted(files.keys()):
-            abs_paths = self.find_files(*key, step=step, index=index)
-
-            if not isinstance(abs_paths, (list, tuple, set)):
-                abs_paths = [abs_paths]
-
-            abs_paths = zip(abs_paths, files[(key, step, index)])
-            abs_paths = sorted(abs_paths, key=lambda p: p[0])
-
-            for abs_path, value in abs_paths:
-                if not abs_path:
-                    raise FileNotFoundError(f"{value.get()} could not be copied")
-
-                if abs_path.startswith(directory):
-                    # File already imported in directory
-                    continue
-
-                import_path = os.path.join(directory, value.get_hashed_filename())
-                if os.path.exists(import_path):
-                    continue
-
-                if verbose:
-                    self.logger.info(f"  Collecting file: {abs_path}")
-                shutil.copy2(abs_path, import_path)
 
     def history(self, job: str) -> "Project":
         '''
