+++ tests/test_project.py
@@ -222,11 +222,6 @@ def test_getworkdir_step_index():
         os.path.abspath(os.path.join("build", "testname", "job0", "thisstep", "thisindex"))
 
 
-def test_getcollectiondir():
-    assert Project("testname").getcollectiondir() == \
-        os.path.abspath(os.path.join("build", "testname", "job0", "sc_collected_files"))
-
-
 def test_record_history():
     proj = Project("testname")
     assert proj.getkeys("history") == tuple()
@@ -944,270 +939,6 @@ def test_summary_select_unknownjob(monkeypatch, caplog):
 
         history.assert_called_once_with("thatjob")  # will call with first alphabetical job
     assert "job0 not found in history, picking thatjob" in caplog.text
-
-
-def test_collect_file_verbose(monkeypatch, caplog):
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_file("top.v")
-    with open("top.v", "w") as f:
-        f.write("test")
-
-    proj = Project(design)
-    monkeypatch.setattr(proj, "_Project__logger", logging.getLogger())
-    proj.logger.setLevel(logging.INFO)
-
-    proj.collect()
-
-    assert f"Collecting files to: {proj.getcollectiondir()}" in caplog.text
-    assert f"  Collecting file: {os.path.abspath('top.v')}" in caplog.text
-
-
-def test_collect_file_not_verbose(monkeypatch, caplog):
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_file("top.v")
-    with open("top.v", "w") as f:
-        f.write("test")
-
-    proj = Project(design)
-    monkeypatch.setattr(proj, "_Project__logger", logging.getLogger())
-    proj.logger.setLevel(logging.INFO)
-
-    proj.collect(verbose=False)
-
-    assert caplog.text == ""
-
-
-def test_collect_file_update():
-    # Checks if collected files are properly updated after editing
-
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_file("fake.v")
-
-    # Edit file
-    with open('fake.v', 'w') as f:
-        f.write('fake')
-
-    proj = Project(design)
-    proj.collect()
-
-    filename = design.get_file(fileset="rtl", filetype="verilog")[0]
-
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-    with open(os.path.join(proj.getcollectiondir(), os.path.basename(filename)), 'r') as f:
-        assert f.readline() == 'fake'
-
-    # Edit file
-    with open('fake.v', 'w') as f:
-        f.write('newfake')
-
-    # Rerun collect
-    proj.collect()
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-    with open(os.path.join(proj.getcollectiondir(), os.path.basename(filename)), 'r') as f:
-        assert f.readline() == 'newfake'
-
-
-def test_collect_directory():
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir("testingdir")
-            design.add_file("testingdir/test.v")
-
-    os.makedirs('testingdir', exist_ok=True)
-
-    with open('testingdir/test.v', 'w') as f:
-        f.write('test')
-
-    proj = Project(design)
-    proj.collect()
-
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-
-    path = design.get_idir(fileset="rtl")[0]
-    assert path.startswith(proj.getcollectiondir())
-    assert os.listdir(path) == ['test.v']
-    assert design.get_file(fileset="rtl",
-                           filetype="verilog")[0].startswith(proj.getcollectiondir())
-
-
-def test_collect_subdirectory():
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir("testingdir")
-            design.add_file("testingdir/subdir/test.v")
-
-    os.makedirs('testingdir/subdir', exist_ok=True)
-
-    with open('testingdir/subdir/test.v', 'w') as f:
-        f.write('test')
-
-    proj = Project(design)
-    proj.collect()
-
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-
-    path = design.get_idir(fileset="rtl")[0]
-    assert path.startswith(proj.getcollectiondir())
-    assert os.listdir(path) == ['subdir']
-    assert os.listdir(os.path.join(path, "subdir")) == ['test.v']
-    assert design.get_file(fileset="rtl",
-                           filetype="verilog")[0].startswith(proj.getcollectiondir())
-
-
-def test_collect_file_with_false():
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=False):
-            design.add_file("fake.v")
-
-    # Edit file
-    with open('fake.v', 'w') as f:
-        f.write('fake')
-
-    proj = Project(design)
-    proj.collect()
-
-    # No files should have been collected
-    assert len(os.listdir(proj.getcollectiondir())) == 0
-
-
-def test_collect_file_home(monkeypatch):
-    def _mock_home():
-        return Path(os.getcwd()) / "home"
-
-    monkeypatch.setattr(Path, 'home', _mock_home)
-
-    _mock_home().mkdir(exist_ok=True)
-
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir(str(Path.home()))
-
-    with open(Path.home() / "test.v", "w") as f:
-        f.write("test")
-
-    proj = Project(design)
-    proj.collect()
-
-    # No files should have been collected
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-    subdir = os.path.join(proj.getcollectiondir(), os.listdir(proj.getcollectiondir())[0])
-    assert len(os.listdir(subdir)) == 0
-
-
-def test_collect_file_build():
-    os.makedirs('build', exist_ok=True)
-
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir("build")
-
-    with open("build/test.v", "w") as f:
-        f.write("test")
-
-    proj = Project(design)
-    proj.collect()
-
-    # No files should have been collected
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-    subdir = os.path.join(proj.getcollectiondir(), os.listdir(proj.getcollectiondir())[0])
-    assert len(os.listdir(subdir)) == 0
-
-
-def test_collect_file_hidden_dir():
-    os.makedirs('test/.test', exist_ok=True)
-
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir("test")
-
-    with open("test/.test/test.v", "w") as f:
-        f.write("test")
-
-    proj = Project(design)
-    proj.collect()
-
-    # No files should have been collected
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-    subdir = os.path.join(proj.getcollectiondir(), os.listdir(proj.getcollectiondir())[0])
-    assert len(os.listdir(subdir)) == 0
-
-
-def test_collect_file_hidden_file():
-    os.makedirs('test', exist_ok=True)
-
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir("test")
-
-    with open("test/.test.v", "w") as f:
-        f.write("test")
-
-    proj = Project(design)
-    proj.collect()
-
-    # No files should have been collected
-    assert len(os.listdir(proj.getcollectiondir())) == 1
-    subdir = os.path.join(proj.getcollectiondir(), os.listdir(proj.getcollectiondir())[0])
-    assert len(os.listdir(subdir)) == 0
-
-
-def test_collect_file_whitelist_error():
-    os.makedirs('test/testing', exist_ok=True)
-
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir("test")
-
-    with open('test/test', 'w') as f:
-        f.write('test')
-
-    proj = Project(design)
-
-    with pytest.raises(RuntimeError, match=".* is not on the approved collection list"):
-        proj.collect(whitelist=[os.path.abspath('not_test_folder')])
-
-    assert len(os.listdir(proj.getcollectiondir())) == 0
-
-
-def test_collect_file_whitelist_pass():
-    os.makedirs('test/testing', exist_ok=True)
-
-    # Create instance of design
-    design = Design("testdesign")
-    with design.active_fileset("rtl"):
-        with design._active(copy=True):
-            design.add_idir("test")
-
-    with open('test/test', 'w') as f:
-        f.write('test')
-
-    proj = Project(design)
-    proj.collect(whitelist=[os.path.abspath('test')])
-
-    assert len(os.listdir(proj.getcollectiondir())) == 1
 
 
 def test_get_task():
